backend/requirements.txt
# --- Các thư viện cũ của bạn ---
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymongo
python-dotenv
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator
fastapi-mail
websockets
yfinance
pandas

# --- CÁC THƯ VIỆN CẦN BỔ SUNG (BẮT BUỘC) ---
newspaper3k          # Để cào báo
google-generativeai  # Để gọi Gemini AI
apscheduler          # Để lên lịch chạy 12 tiếng/lần
requests             # Thư viện hỗ trợ mạng
lxml_html_clean   # Hỗ trợ xử lý HTML cho newspaper3k
tenacity

# ---- Thi vien chatbot ------------------
neo4j
langchain-community
langchain-openai